requirements.txt
torch
torchvision
numpy
tqdm
Pillow
opencv-python
# scikit-learn  # add only if you actually use it in reports